add SpecificInformation messages handling
handle MSG_HINT (requires work still)
handle MSG_WIN
handle MSG_MATCH_KILL

the match kill message actually doesnt return any message,
instead it sets the state of the object in such a way
that when the win message is received, that information is
passed away to the clients. A way to check if the `Encode`
function is returning a message worth sending is needed.